Cambiar el tipo de retorno del método tiempoLlegada a double para mejorar la precisión del cálculo.

+++ actividad02/ejercicio03/Automovil.java
@@ -119,7 +119,7 @@ public class Automovil {
 
     }
 
-    int tiempoLlegada(int distancia) {
+    double tiempoLlegada(int distancia) {
         if (velocidadActual > 0) {
             return distancia / velocidadActual;
         } else {
